+++ CHANGES.md
@@ -20,6 +20,7 @@ This file is a best-effort approach to solving this issue; we will do our best b
 * (wifi) Added new attributes to `AdhocWifiMac` to control beacon transmissions: **BeaconGeneration** and **BeaconInterval** with the same meaning as for `ApWifiMac`. It is also possible to control the access category to use for beacons via the **BeaconAc** attribute.
 * (internet) Added a new helper (P2pCacheHelper) to set up auto-generated ARP cache for P2P devices.
 * (wifi) Added a new attribute `Per20CcaSensitivityThreshold` to `EhtConfiguration` for tuning the Per 20MHz CCA threshold when 802.11be is used.
+* (internet) A new trace source `Ipv6L3Protocol::MulticastForward` has been added with the same meaning as for IPv4.
 
 ### Changes to existing API